requirements.txt
altair==4.1.0
ansi2html==1.8.0
anyio==3.7.1
appnope==0.1.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
astor==0.8.1
asttokens==2.2.1
async-lru==2.0.3
attrs==23.1.0
Babel==2.12.1
backcall==0.2.0
base58==2.1.1
beautifulsoup4==4.12.2
bleach==6.0.0
blinker==1.5
boto3==1.25.1
botocore==1.28.1
cachelib==0.9.0
cachetools==5.2.0
certifi==2020.12.5
cffi==1.15.1
chardet==4.0.0
charset-normalizer==2.0.4
click==8.1.3
colorama==0.4.6
comm==0.1.3
commonmark==0.9.1
dash==2.11.1
dash-core-components==2.0.0
dash-cytoscape==0.3.0
dash-extensions==1.0.1
dash-html-components==2.0.0
dash-table==5.0.0
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
distro==1.8.0
EditorConfig==0.12.3
entrypoints==0.4
exceptiongroup==1.1.2
executing==1.2.0
fastjsonschema==2.18.0
Flask==2.2.5
Flask-Caching==2.0.2
fqdn==1.5.1
gitdb==4.0.9
GitPython==3.1.29
idna==2.10
importlib-metadata==5.0.0
ipykernel==6.25.0
ipython==8.14.0
ipython-genutils==0.2.0
ipywidgets==8.0.7
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.18.2
jellyfish==0.6.1
Jinja2==3.1.2
jmespath==1.0.1
jsbeautifier==1.14.9
json5==0.9.14
jsonpointer==2.4
jsonschema==4.18.4
jsonschema-specifications==2023.7.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.6.3
jupyter-lsp==2.2.0
jupyter_client==8.3.0
jupyter_core==5.3.1
jupyter_server==2.7.0
jupyter_server_terminals==0.4.4
jupyterlab==4.0.3
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.8
jupyterlab_server==2.24.0
lxml==4.6.3
MarkupSafe==2.1.1
matplotlib-inline==0.1.6
mistune==3.0.1
more-itertools==9.1.0
multitasking==0.0.9
nbclient==0.8.0
nbconvert==7.7.3
nbformat==5.9.1
nest-asyncio==1.5.6
nodejs==0.1.1
notebook==7.0.0
notebook_shim==0.2.3
npm==0.1.1
numpy==1.19.5
numpy-financial==1.0.0
opencv-python==4.5.2.54
optional-django==0.1.0
overrides==7.3.1
packaging==21.3
pandas==1.1.5
pandas-datareader==0.9.0
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
platformdirs==3.9.1
plotly==5.15.0
prometheus-client==0.17.1
prompt-toolkit==3.0.39
protobuf==3.15.5
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
py-cn-phone-area-code==1.0.5
py-postgresql==1.2.2
pyarrow==9.0.0
pyathena==2.2.0
pycparser==2.21
pycryptodome==3.10.1
pycryptodomex==3.10.1
pydeck==0.8.0b4
Pygments==2.13.0
PyJWT==2.0.1
Pympler==1.0.1
pyparsing==3.0.9
pypng==0.20220715.0
pyrsistent==0.18.1
python-dateutil==2.8.2
python-json-logger==2.0.7
pytz==2021.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0.1
pyzmq==25.1.0
qrcode==7.4.2
qtconsole==5.4.3
QtPy==2.3.1
referencing==0.30.0
requests==2.31.0
retrying==1.3.4
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==12.6.0
rpds-py==0.9.2
s3transfer==0.6.0
semver==2.13.0
Send2Trash==1.8.2
six==1.15.0
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.4.1
stack-data==0.6.2
streamlit==1.15.1
tabula-py==2.7.0
tenacity==8.1.0
terminado==0.17.1
tinycss2==1.2.1
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
tornado==6.2
traitlets==5.9.0
typing_extensions==4.7.1
tzdata==2022.5
tzlocal==4.2
uri-template==1.3.0
urllib3==1.26.4
us==2.0.2
validators==0.20.0
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.1
Werkzeug==2.2.3
widgetsnbextension==4.0.8
XlsxWriter==1.2.8
yfinance==0.1.59
zipp==3.10.0
